+++ stackslib/src/chainstate/stacks/tests/chain_histories.rs
@@ -2734,7 +2734,12 @@ pub fn mine_empty_anchored_block(
     let tx_coinbase_signed = make_coinbase(miner, burnchain_height);
 
     builder
-        .try_mine_tx(clarity_tx, &tx_coinbase_signed, ASTRules::PrecheckSize)
+        .try_mine_tx(
+            clarity_tx,
+            &tx_coinbase_signed,
+            ASTRules::PrecheckSize,
+            None,
+        )
         .unwrap();
 
     let stacks_block = builder.mine_anchored_block(clarity_tx);
@@ -2769,7 +2774,12 @@ pub fn mine_empty_anchored_block_with_burn_height_pubkh(
     let tx_coinbase_signed = make_coinbase(miner, burnchain_height);
 
     builder
-        .try_mine_tx(clarity_tx, &tx_coinbase_signed, ASTRules::PrecheckSize)
+        .try_mine_tx(
+            clarity_tx,
+            &tx_coinbase_signed,
+            ASTRules::PrecheckSize,
+            None,
+        )
         .unwrap();
 
     let stacks_block = builder.mine_anchored_block(clarity_tx);
@@ -2804,7 +2814,12 @@ pub fn mine_empty_anchored_block_with_stacks_height_pubkh(
     let tx_coinbase_signed = make_coinbase(miner, burnchain_height);
 
     builder
-        .try_mine_tx(clarity_tx, &tx_coinbase_signed, ASTRules::PrecheckSize)
+        .try_mine_tx(
+            clarity_tx,
+            &tx_coinbase_signed,
+            ASTRules::PrecheckSize,
+            None,
+        )
         .unwrap();
 
     let stacks_block = builder.mine_anchored_block(clarity_tx);
@@ -2835,7 +2850,12 @@ pub fn mine_invalid_token_transfers_block(
     // make a coinbase for this miner
     let tx_coinbase_signed = make_coinbase(miner, burnchain_height);
     builder
-        .try_mine_tx(clarity_tx, &tx_coinbase_signed, ASTRules::PrecheckSize)
+        .try_mine_tx(
+            clarity_tx,
+            &tx_coinbase_signed,
+            ASTRules::PrecheckSize,
+            None,
+        )
         .unwrap();
 
     let recipient =
@@ -2909,7 +2929,12 @@ pub fn mine_smart_contract_contract_call_block(
     // make a coinbase for this miner
     let tx_coinbase_signed = make_coinbase(miner, burnchain_height);
     builder
-        .try_mine_tx(clarity_tx, &tx_coinbase_signed, ASTRules::PrecheckSize)
+        .try_mine_tx(
+            clarity_tx,
+            &tx_coinbase_signed,
+            ASTRules::PrecheckSize,
+            None,
+        )
         .unwrap();
 
     // make a smart contract
@@ -2919,7 +2944,12 @@ pub fn mine_smart_contract_contract_call_block(
         builder.header.total_work.work as usize,
     );
     builder
-        .try_mine_tx(clarity_tx, &tx_contract_signed, ASTRules::PrecheckSize)
+        .try_mine_tx(
+            clarity_tx,
+            &tx_contract_signed,
+            ASTRules::PrecheckSize,
+            None,
+        )
         .unwrap();
 
     // make a contract call
@@ -2931,7 +2961,12 @@ pub fn mine_smart_contract_contract_call_block(
         2,
     );
     builder
-        .try_mine_tx(clarity_tx, &tx_contract_call_signed, ASTRules::PrecheckSize)
+        .try_mine_tx(
+            clarity_tx,
+            &tx_contract_call_signed,
+            ASTRules::PrecheckSize,
+            None,
+        )
         .unwrap();
 
     let stacks_block = builder.mine_anchored_block(clarity_tx);
@@ -2986,7 +3021,12 @@ pub fn mine_smart_contract_block_contract_call_microblock(
     // make a coinbase for this miner
     let tx_coinbase_signed = make_coinbase(miner, burnchain_height);
     builder
-        .try_mine_tx(clarity_tx, &tx_coinbase_signed, ASTRules::PrecheckSize)
+        .try_mine_tx(
+            clarity_tx,
+            &tx_coinbase_signed,
+            ASTRules::PrecheckSize,
+            None,
+        )
         .unwrap();
 
     // make a smart contract
@@ -2996,7 +3036,12 @@ pub fn mine_smart_contract_block_contract_call_microblock(
         builder.header.total_work.work as usize,
     );
     builder
-        .try_mine_tx(clarity_tx, &tx_contract_signed, ASTRules::PrecheckSize)
+        .try_mine_tx(
+            clarity_tx,
+            &tx_contract_signed,
+            ASTRules::PrecheckSize,
+            None,
+        )
         .unwrap();
 
     let stacks_block = builder.mine_anchored_block(clarity_tx);
@@ -3073,7 +3118,12 @@ pub fn mine_smart_contract_block_contract_call_microblock_exception(
     // make a coinbase for this miner
     let tx_coinbase_signed = make_coinbase(miner, burnchain_height);
     builder
-        .try_mine_tx(clarity_tx, &tx_coinbase_signed, ASTRules::PrecheckSize)
+        .try_mine_tx(
+            clarity_tx,
+            &tx_coinbase_signed,
+            ASTRules::PrecheckSize,
+            None,
+        )
         .unwrap();
 
     // make a smart contract
@@ -3083,7 +3133,12 @@ pub fn mine_smart_contract_block_contract_call_microblock_exception(
         builder.header.total_work.work as usize,
     );
     builder
-        .try_mine_tx(clarity_tx, &tx_contract_signed, ASTRules::PrecheckSize)
+        .try_mine_tx(
+            clarity_tx,
+            &tx_contract_signed,
+            ASTRules::PrecheckSize,
+            None,
+        )
         .unwrap();
 
     let stacks_block = builder.mine_anchored_block(clarity_tx);