+++ stackslib/src/net/mod.rs
@@ -3097,6 +3097,7 @@ pub mod test {
                             &boot_code_smart_contract,
                             &boot_code_account,
                             ASTRules::PrecheckSize,
+                            None,
                         )
                         .unwrap()
                     });